chore: moving selinux policies to utils
